backend/requirements.txt
# Django and core dependencies
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.10

# Environment variables
python-dotenv==1.0.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.2
fpdf2==2.7.6

# Cloudinary storage
django-cloudinary-storage==0.3.0
cloudinary>=1.36.0

# PDF rendering (HTML to PDF)
jinja2==3.1.4
weasyprint>=67.0
pydyf>=0.12.0

# HTTP requests
requests==2.32.3

# Production server
gunicorn==23.0.0

# Static file serving
whitenoise==6.8.2

#JWT
PyJWT>=2.7.0

# API documentation (Swagger/OpenAPI)
drf-spectacular==0.27.2

# Vector search
pgvector>=0.3.6
openai>=1.0.0
